basically I think I solve the layout problem
however when I set the check box of subclass filter to false in program.

the level information is not correct, means the tree info is not
generated correctly, because the arcs are not added to the model.

I still haven't look in to the code to find the logic of the visibility
of an arc.

The comon_relTest version should be used to further implement the comon
ontology