tree explorer expand / collapse can control the graph correctly right now. :）
It would be better that graph can only affect the selection of the tree
explorer, and show the explanation as both tooltip on the node as well
as text right bellow the tree explorer lable.

If graph expansion / collapse can control the tree explorer, it might
have trouble in sync both sides, and may be too flexible.

Tree explorer should have tool icon to full expand the selected node, or
to hide / show the out most layer( that can not be called expand/
collapse).